set authentication and authorization for deleting a post

+++ controllers/api/v1/posts_api.js
@@ -1,5 +1,5 @@
 const Post = require("../../../models/post");
-const Comment=require('../../../models/comment');
+const Comment = require("../../../models/comment");
 
 module.exports.index = async function (req, res) {
   try {
@@ -33,19 +33,19 @@ module.exports.destroy = async function (req, res) {
     const post = await Post.findById(req.params.id);
 
     // .id means converting the objectId into string
-    // if (post && post.user == req.user.id) {
-    await Post.deleteOne({ _id: req.params.id });
+    if (post && post.user == req.user.id) {
+      await Post.deleteOne({ _id: req.params.id });
 
-    await Comment.deleteMany({ post: req.params.id });
+      await Comment.deleteMany({ post: req.params.id });
 
-    return res.status(200).json({
-      message: "Post and associated comments deleted!",
+      return res.status(200).json({
+        message: "Post and associated comments deleted!",
+      });
+    }
+
+    return res.status(401).json({
+      message: "You are not authorized to delete this post!",
     });
-    // }
-
-    // return res.status(401).json({
-    //   message: "You are not authorized to delete this post!",
-    // });
   } catch (err) {
     return res.status(500).json({
       message: "Internal Server Error",
